--- a/lock.ts
+++ b/lock.ts
@@ -43,7 +43,11 @@ export class Lock<T> {
    * @returns A Promise that resolves with the result of the function.
    */
   async lock<R>(fn: (value: T) => R | PromiseLike<R>): Promise<R> {
-    using _lock = await this.#mu.acquire();
-    return await fn(this.#value);
+    const lock = await this.#mu.acquire();
+    try {
+      return await fn(this.#value);
+    } finally {
+      lock[Symbol.dispose]();
+    }
   }
 }