--- a/semaphore.ts
+++ b/semaphore.ts
@@ -1,3 +1,5 @@
+import { RawSemaphore } from "./_raw_semaphore.ts";
+
 /**
  * A semaphore that allows a limited number of concurrent executions of an operation.
  *
@@ -14,9 +16,7 @@
  * ```
  */
 export class Semaphore {
-  #resolves: (() => void)[] = [];
-  #value: number;
-  #size: number;
+  #sem: RawSemaphore;
 
   /**
    * Creates a new semaphore with the specified limit.
@@ -25,20 +25,14 @@ export class Semaphore {
    * @throws {RangeError} if the size is not a positive safe integer.
    */
   constructor(size: number) {
-    if (size <= 0 || !Number.isSafeInteger(size)) {
-      throw new RangeError(
-        `size must be a positive safe integer, got ${size}`,
-      );
-    }
-    this.#value = size;
-    this.#size = size;
+    this.#sem = new RawSemaphore(size);
   }
 
   /**
    * Returns true if the semaphore is currently locked.
    */
   get locked(): boolean {
-    return this.#value === 0;
+    return this.#sem.locked;
   }
 
   /**
@@ -48,31 +42,11 @@ export class Semaphore {
    * @returns A promise that resolves to the return value of the specified function.
    */
   async lock<R>(fn: () => R | PromiseLike<R>): Promise<R> {
-    await this.#acquire();
+    await this.#sem.acquire();
     try {
       return await fn();
     } finally {
-      this.#release();
-    }
-  }
-
-  #acquire(): Promise<void> {
-    if (this.#value > 0) {
-      this.#value -= 1;
-      return Promise.resolve();
-    } else {
-      const { promise, resolve } = Promise.withResolvers<void>();
-      this.#resolves.push(resolve);
-      return promise;
-    }
-  }
-
-  #release(): void {
-    const resolve = this.#resolves.shift();
-    if (resolve) {
-      resolve();
-    } else if (this.#value < this.#size) {
-      this.#value += 1;
+      this.#sem.release();
     }
   }
 }