Continue from #31 

```
cpu: Apple M1 Max
runtime: deno 1.45.5 (aarch64-apple-darwin)

file:///Users/alisue/ogh/jsr-core/asyncutil/barrier_bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------

group Barrier#wait
current                    184.58 µs/iter       5,417.8                            (165.38 µs … 517.5 µs)              173.17 µs              419.12 µs              431.88 µs
v1.0.0                     248.35 µs/iter       4,026.5                              (205 µs … 565.21 µs)                 229 µs              510.62 µs              517.75 µs

summary
  current
   1.35x faster than v1.0.0


file:///Users/alisue/ogh/jsr-core/asyncutil/lock_bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------

group Lock#lock
current                    238.51 µs/iter       4,192.6                           (211.46 µs … 549.62 µs)              225.79 µs              496.46 µs               507.5 µs
v1.0.0                       52.9 ms/iter          18.9                             (47.09 ms … 56.71 ms)               55.55 ms               56.71 ms               56.71 ms

summary
  current
   221.8x faster than v1.0.0


file:///Users/alisue/ogh/jsr-core/asyncutil/mutex_bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------

group Mutex#wait
current                    414.06 µs/iter       2,415.1                              (366.67 µs … 2.8 ms)              384.33 µs              757.71 µs                1.65 ms
v1.0.0                      52.04 ms/iter          19.2                             (46.98 ms … 56.89 ms)               54.69 ms               56.89 ms               56.89 ms

summary
  current
   125.68x faster than v1.0.0


file:///Users/alisue/ogh/jsr-core/asyncutil/notify_bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------

group Notify#notifyAll
current                     296.2 µs/iter       3,376.1                            (252.5 µs … 730.54 µs)              269.75 µs              603.04 µs              611.88 µs
v1.0.0                     230.36 µs/iter       4,341.1                           (199.62 µs … 548.25 µs)              223.17 µs              409.88 µs              417.46 µs

summary
  current
   1.29x slower than v1.0.0

group Notify#notify
current                     406.6 µs/iter       2,459.4                           (345.54 µs … 831.54 µs)              379.38 µs              763.67 µs              790.67 µs
v1.0.0                      22.33 ms/iter          44.8                             (21.08 ms … 37.03 ms)               22.09 ms               37.03 ms               37.03 ms

summary
  current
   54.92x faster than v1.0.0


file:///Users/alisue/ogh/jsr-core/asyncutil/queue_bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------

group Queue#push/pop
current                     154.2 µs/iter       6,485.2                           (146.21 µs … 329.67 µs)              151.42 µs              263.92 µs              275.79 µs
v1.0.0                       1.13 ms/iter         881.3                             (919.08 µs … 5.15 ms)                1.01 ms                3.97 ms                4.73 ms

summary
  current
   7.36x faster than v1.0.0


file:///Users/alisue/ogh/jsr-core/asyncutil/rw_lock_bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------

group RwLock#lock
current                    271.94 µs/iter       3,677.2                              (247.58 µs … 689 µs)              258.33 µs              519.33 µs              540.79 µs
v1.0.0                      54.55 ms/iter          18.3                             (49.02 ms … 61.45 ms)               56.62 ms               61.45 ms               61.45 ms

summary
  current
   200.59x faster than v1.0.0

group RwLock#rlock
current                    206.41 µs/iter       4,844.7                           (182.83 µs … 558.88 µs)              196.38 µs              457.38 µs              478.71 µs
v1.0.0                      55.64 ms/iter          18.0                             (50.13 ms … 62.36 ms)               57.77 ms               62.36 ms               62.36 ms

summary
  current
   269.58x faster than v1.0.0


file:///Users/alisue/ogh/jsr-core/asyncutil/semaphore_bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------

group Semaphore#lock
current                    231.89 µs/iter       4,312.4                           (207.62 µs … 547.54 µs)              215.67 µs              485.46 µs              500.75 µs
v1.0.0                      21.67 ms/iter          46.2                             (20.63 ms … 24.15 ms)               21.93 ms               24.15 ms               24.15 ms

summary
  current
   93.43x faster than v1.0.0


file:///Users/alisue/ogh/jsr-core/asyncutil/stack_bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------

group Stack#push/pop
current                     111.6 µs/iter       8,960.5                           (100.17 µs … 497.96 µs)              110.21 µs              215.75 µs              223.58 µs
v1.0.0                       1.11 ms/iter         903.9                             (882.92 µs … 6.59 ms)                 996 µs                3.56 ms                4.91 ms

summary
  current
   9.91x faster than v1.0.0


file:///Users/alisue/ogh/jsr-core/asyncutil/wait_group_bench.ts
benchmark      time (avg)        iter/s             (min … max)       p75       p99      p995
--------------------------------------------------------------- -----------------------------

group WaitGroup#wait
current                     64.17 µs/iter      15,583.6                             (59.5 µs … 209.42 µs)               66.88 µs               79.04 µs               92.62 µs
v1.0.0                      64.23 µs/iter      15,568.3                             (59.5 µs … 311.67 µs)               67.08 µs               78.92 µs                88.5 µs

summary
  current
   1x faster than v1.0.0
```